Rename points.js to award-points.js
deductPoints now moved to own file
